--- host-contracts/tasks/utils/kmsGenerationMigrationEnvKeys.ts
@@ -1,21 +1,0 @@
-export const KMS_GENERATION_MIGRATION_ENV_KEYS = [
-  'MIGRATION_PREP_KEYGEN_COUNTER',
-  'MIGRATION_KEY_COUNTER',
-  'MIGRATION_CRS_COUNTER',
-  'MIGRATION_ACTIVE_KEY_ID',
-  'MIGRATION_ACTIVE_CRS_ID',
-  'MIGRATION_ACTIVE_PREP_KEYGEN_ID',
-  'MIGRATION_ACTIVE_KEY_DIGESTS',
-  'MIGRATION_ACTIVE_CRS_DIGEST',
-  'MIGRATION_KEY_CONSENSUS_TX_SENDERS',
-  'MIGRATION_KEY_CONSENSUS_DIGEST',
-  'MIGRATION_CRS_CONSENSUS_TX_SENDERS',
-  'MIGRATION_CRS_CONSENSUS_DIGEST',
-  'MIGRATION_PREP_KEYGEN_CONSENSUS_TX_SENDERS',
-  'MIGRATION_PREP_KEYGEN_CONSENSUS_DIGEST',
-  'MIGRATION_CRS_MAX_BIT_LENGTH',
-  'MIGRATION_PREP_KEYGEN_PARAMS_TYPE',
-  'MIGRATION_CRS_PARAMS_TYPE',
-  'MIGRATION_CONTEXT_ID',
-] as const;
-
